typedef enum

+++ Ch10_Example/Ch10_Typedef_Enum_Example.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-typedef double SCALAR;
+typedef float SCALAR;
 
 enum colors
 {
